Restructure project and add scripts
