+++ scripts/train_g2s.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+python train/train.py config/smile_config.yaml
