+++ scripts/validate.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+python validate.py --config config/smile_config.yaml
