+++ train/train.py
@@ -4,7 +4,7 @@ import torch
 import torch.nn as nn
 import torch.optim as optim
 from torch.utils.data import DataLoader
-from transformer import TransformerModel
+from models.transformer import TransformerModel
 from utils import ReactionSequenceDataset, train, validation
 import sys
 import yaml
